# backend/internal/handler/learning_handler.go
package handler

import (
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/virtualguard/PaperBeginer/pkg/response"
)

// GenerateLearningPath generates a learning path
// @Summary Generate learning path
// @Description Generate an AI-powered learning path for a research field
// @Tags learning
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body GenerateLearningPathRequest true "Generation request"
// @Success 202 {object} response.Response
// @Failure 400 {object} response.Response
// @Failure 401 {object} response.Response
// @Router /learning/generate [post]
func GenerateLearningPath(c *gin.Context) {
	var req GenerateLearningPathRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.BadRequest(c, "Invalid request body")
		return
	}

	// TODO: Implement learning path generation
	response.Success(c, gin.H{
		"message":     "Learning path generation queued",
		"category_id": req.CategoryID,
		"difficulty":  req.Difficulty,
	})
}

// GetLearningPaths gets user's learning paths
// @Summary Get learning paths
// @Description Get user's generated learning paths
// @Tags learning
// @Produce json
// @Security BearerAuth
// @Success 200 {object} response.Response
// @Failure 401 {object} response.Response
// @Router /learning/paths [get]
func GetLearningPaths(c *gin.Context) {
	// TODO: Implement learning paths retrieval
	response.Success(c, gin.H{
		"paths":   []interface{}{},
		"message": "Learning paths feature coming soon",
	})
}

// GetLearningPath gets a specific learning path
// @Summary Get learning path
// @Description Get a specific learning path by ID
// @Tags learning
// @Produce json
// @Security BearerAuth
// @Param id path string true "Learning path ID"
// @Success 200 {object} response.Response
// @Failure 401 {object} response.Response
// @Failure 404 {object} response.Response
// @Router /learning/paths/{id} [get]
func GetLearningPath(c *gin.Context) {
	pathID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.BadRequest(c, "Invalid learning path ID")
		return
	}

	// TODO: Implement learning path retrieval
	_ = pathID

	response.NotFound(c, "Learning path not found")
}

// GenerateLearningPathRequest represents a learning path generation request
type GenerateLearningPathRequest struct {
	CategoryID    int    `json:"category_id" binding:"required"`
	Difficulty    string `json:"difficulty" binding:"required,oneof=beginner intermediate advanced"`
	Prerequisites []string `json:"prerequisites,omitempty"`
	FocusAreas    []string `json:"focus_areas,omitempty"`
}