# backend/internal/handler/review_handler.go
package handler

import (
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/virtualguard/PaperBeginer/internal/handler/middleware"
	"github.com/virtualguard/PaperBeginer/pkg/response"
)

// GenerateReview generates a literature review
// @Summary Generate review
// @Description Generate an AI-powered literature review from papers
// @Tags reviews
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body GenerateReviewRequest true "Generation request"
// @Success 202 {object} response.Response
// @Failure 400 {object} response.Response
// @Failure 401 {object} response.Response
// @Router /reviews/generate [post]
func GenerateReview(c *gin.Context) {
	userID, err := middleware.GetUserID(c)
	if err != nil {
		response.Unauthorized(c, "Invalid token")
		return
	}

	var req GenerateReviewRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.BadRequest(c, "Invalid request body")
		return
	}

	if len(req.PaperIDs) == 0 {
		response.BadRequest(c, "At least one paper ID is required")
		return
	}

	// TODO: Implement review generation
	_ = userID

	response.Success(c, gin.H{
		"message":   "Review generation queued",
		"paper_ids": req.PaperIDs,
	})
}

// ListReviews lists user's reviews
// @Summary List reviews
// @Description Get a paginated list of user's reviews
// @Tags reviews
// @Produce json
// @Security BearerAuth
// @Param page query int false "Page number" default(1)
// @Param per_page query int false "Items per page" default(20)
// @Success 200 {object} response.Response
// @Failure 401 {object} response.Response
// @Router /reviews [get]
func ListReviews(c *gin.Context) {
	userID, err := middleware.GetUserID(c)
	if err != nil {
		response.Unauthorized(c, "Invalid token")
		return
	}

	page, _ := strconv.Atoi(c.DefaultQuery("page", "1"))
	perPage, _ := strconv.Atoi(c.DefaultQuery("per_page", "20"))

	// TODO: Implement reviews listing
	_ = userID
	_ = page
	_ = perPage

	response.Success(c, gin.H{
		"reviews": []interface{}{},
		"message": "Reviews feature coming soon",
	})
}

// GetReview gets a specific review
// @Summary Get review
// @Description Get a specific review by ID
// @Tags reviews
// @Produce json
// @Security BearerAuth
// @Param id path string true "Review ID"
// @Success 200 {object} response.Response
// @Failure 401 {object} response.Response
// @Failure 404 {object} response.Response
// @Router /reviews/{id} [get]
func GetReview(c *gin.Context) {
	userID, err := middleware.GetUserID(c)
	if err != nil {
		response.Unauthorized(c, "Invalid token")
		return
	}

	reviewID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.BadRequest(c, "Invalid review ID")
		return
	}

	// TODO: Implement review retrieval
	_ = userID
	_ = reviewID

	response.NotFound(c, "Review not found")
}

// ScoreReview scores a user's review
// @Summary Score review
// @Description Get AI scoring for a user-written review
// @Tags reviews
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Review ID"
// @Param request body ScoreReviewRequest true "Scoring request"
// @Success 200 {object} response.Response
// @Failure 400 {object} response.Response
// @Failure 401 {object} response.Response
// @Failure 404 {object} response.Response
// @Router /reviews/{id}/score [post]
func ScoreReview(c *gin.Context) {
	userID, err := middleware.GetUserID(c)
	if err != nil {
		response.Unauthorized(c, "Invalid token")
		return
	}

	reviewID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.BadRequest(c, "Invalid review ID")
		return
	}

	var req ScoreReviewRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		// Content is optional if we're scoring an existing review
	}

	// TODO: Implement review scoring
	_ = userID
	_ = reviewID
	_ = req

	response.Success(c, gin.H{
		"message": "Review scoring queued",
	})
}

// DeleteReview deletes a review
// @Summary Delete review
// @Description Delete a review by ID
// @Tags reviews
// @Produce json
// @Security BearerAuth
// @Param id path string true "Review ID"
// @Success 204 "No Content"
// @Failure 401 {object} response.Response
// @Failure 404 {object} response.Response
// @Router /reviews/{id} [delete]
func DeleteReview(c *gin.Context) {
	userID, err := middleware.GetUserID(c)
	if err != nil {
		response.Unauthorized(c, "Invalid token")
		return
	}

	reviewID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.BadRequest(c, "Invalid review ID")
		return
	}

	// TODO: Implement review deletion
	_ = userID
	_ = reviewID

	response.NoContent(c)
}

// GenerateReviewRequest represents a review generation request
type GenerateReviewRequest struct {
	Title      string      `json:"title" binding:"required"`
	PaperIDs   []uuid.UUID `json:"paper_ids" binding:"required"`
	CategoryID *int        `json:"category_id,omitempty"`
	Style      string      `json:"style,omitempty"` // academic, summary, comprehensive
}

// ScoreReviewRequest represents a review scoring request
type ScoreReviewRequest struct {
	Content string `json:"content,omitempty"` // User's review content to score
}